refactored landing page
